+++ packages/emnapi/src/async_context.c
@@ -1,5 +1,5 @@
 #include <node_api.h>
-#include "emnapi_common.h"
+#include "emnapi_internal.h"
 
 EXTERN_C_START
