+++ packages/emnapi/src/async_work.c
@@ -1,5 +1,5 @@
 #include "node_api.h"
-#include "emnapi_common.h"
+#include "emnapi_internal.h"
 
 #if EMNAPI_HAVE_THREADS
