+++ packages/emnapi/src/emnapi.c
@@ -1,4 +1,4 @@
-#include "emnapi_common.h"
+#include "emnapi_internal.h"
 
 EXTERN_C_START
