+++ packages/emnapi/src/emnapi_internal.h
