+++ packages/emnapi/src/js_native_api.c
@@ -1,4 +1,4 @@
-#include "emnapi_common.h"
+#include "emnapi_internal.h"
 
 #ifdef __EMSCRIPTEN__
 #include <emscripten/heap.h>
